scripts/requirements.txt
annotated-types==0.7.0
anyio==4.7.0
attrs==24.2.0
beautifulsoup4==4.12.3
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
config==0.5.1
distro==1.9.0
google-ai-generativelanguage==0.6.10
google-api-core==2.24.0
google-api-python-client==2.159.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
grpcio==1.67.1
grpcio-status==1.67.1
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
jiter==0.8.0
joblib==1.4.2
jsonlines==4.0.0
lxml==5.3.0
numpy==2.2.0
openai==1.57.0
pandas==2.2.3
proto-plus==1.25.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.3
pydantic_core==2.27.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2024.2
RapidFuzz==3.11.0
regex==2024.11.6
requests==2.32.3
rsa==4.9
scikit-learn==1.6.0
scipy==1.14.1
setuptools==75.1.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
threadpoolctl==3.5.0
tiktoken==0.8.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
wheel==0.44.0
